Add Logging support for master merge, Error Handling
